Refactor WebSocket{Protocol}.CreateFromStream tests to be shared
We have the WebSocket.CreateFromStream and WebSocketProtocol.CreateFromStream methods, which are identical, except that the former is netcoreapp-only and the latter is in a separate NuGet package for downlevel use.  However, tests for them were separate, with some tests only for one and some tests for the other.

This commit centralizes those tests so they're shared by and apply to both methods.  There are no code changes to the actual bodies of tests, just moving code around to have it apply to both.
